requirements.txt
# 웹 프레임워크
fastapi==0.115.0
uvicorn==0.32.0

# 환경 변수 관리
python-dotenv==1.0.1

# HTTP 요청
requests==2.32.3

# 데이터 처리
pydantic==2.10.2
numpy>=1.23.0
scikit-learn>=1.2.0

# Google API (YouTube Data API)
google-api-python-client>=2.0.0

# RAG 시스템 의존성
openai>=1.12.0           # OpenAI GPT-4o API
tavily-python>=0.3.0     # Tavily 웹 검색 API
chromadb>=0.4.22         # 벡터 데이터베이스
sentence-transformers>=2.3.0  # 텍스트 임베딩

# 멀티모달 (이미지/오디오) 처리
opencv-python>=4.8.0     # 영상 프레임 추출
yt-dlp>=2023.10.0        # 유튜브 영상 다운로드
easyocr>=1.7.0           # 이미지 텍스트 추출 (OCR)
torch>=2.0.0             # 딥러닝 프레임워크
transformers>=4.30.0     # HuggingFace 모델 (CLIP, Whisper 등)
pillow>=10.0.0           # 이미지 처리
librosa>=0.10.0          # 오디오 분석
soundfile>=0.12.0        # 오디오 파일 처리
openai-whisper>=20231117 # 음성 인식 (자막 추출)


# Google Cloud Vision API
google-cloud-vision>=3.4.0
requests>=2.31.0
dateparser
